+++ docs/changelog.md
@@ -53,6 +53,7 @@ As features stabilize some brief notes about them will accumulate here.
 * New: [wezterm.action_callback](config/lua/wezterm/action_callback.md) function to make it easier to use custom events. Thanks to [@bew](https://github.com/bew)! [#1151](https://github.com/wez/wezterm/pull/1151)
 * New: `wezterm connect` now also supports the `--class` parameter to override the window class
 * Fixed: wezterm can now match bitmap fonts that are spread across multiple font files [#1189](https://github.com/wez/wezterm/issues/1189)
+* Improved: [use_cap_height_to_scale_fallback_fonts](config/lua/config/use_cap_height_to_scale_fallback_fonts.md) now computes *cap-height* based on the rasterized glyph bitmap which means that the data is accurate in more cases, including for bitmap fonts.  Scaling is now also applied to across varying text styles; previously it only applied to a font within an `wezterm.font_with_fallback` font list.
 
 ### 20210814-124438-54e29167